requirements.txt
django >= 1.7, < 1.8
djangorestframework == 3.1.1
django-import-export
psycopg2
nltk == 3.0.2
python-levenshtein == 0.12.0
fuzzywuzzy == 0.5.0

